separated topic to /commands
